Restructuring for heroku
Moved backend files up a directory and made changes so it will run on heroku
