+++ package.json
@@ -1,0 +1,26 @@
+{
+  "name": "heroku",
+  "version": "1.0.0",
+  "main": "server.js",
+  "scripts": {
+    "client-install": "npm install --prefix client",
+	"start": "node server.js",
+	"server": "nodemon server.js",
+	"client": "npm start --prefix client",
+	"dev": "concurrently \"npm run server\" \"npm run client\"",
+	"heroku-postbuild": "NPM_CONFIG_PRODUCTION=false npm install --prefix client && npm run build --prefix client"
+  },
+  "dependencies": {
+    "client": "0.0.1",
+    "concurrently": "^5.3.0",
+    "cookie-parser": "~1.4.4",
+    "cors": "^2.8.5",
+    "create-react-app": "^4.0.0",
+    "debug": "~2.6.9",
+    "express": "~4.16.1",
+    "http-errors": "~1.6.3",
+    "jade": "~1.11.0",
+    "morgan": "~1.9.1",
+    "sentiment": "^5.0.2"
+  }
+}
