+++ routes/searchAPI.js
